Split LRO.py into atomic mixin files
